Adds additional unit tests for `join` method with more than 2 arguments and with Closure as 4-th argument
