+++ tests/Database/Driver/MySQL/SelectWithJoinQueryTest.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * Spiral Framework.
+ *
+ * @license   MIT
+ * @author    Anton Titov (Wolfy-J)
+ */
+
+declare(strict_types=1);
+
+namespace Cycle\Database\Tests\Driver\MySQL;
+
+/**
+ * @group driver
+ * @group driver-mysql
+ */
+class SelectWithJoinQueryTest extends \Cycle\Database\Tests\SelectWithJoinQueryTest
+{
+    const DRIVER = "mysql";
+}
