+++ tests/Database/Driver/Postgres/SelectWithJoinQueryTest.php
@@ -1,0 +1,21 @@
+<?php
+
+/**
+ * Spiral Framework.
+ *
+ * @license   MIT
+ * @author    Anton Titov (Wolfy-J)
+ */
+
+declare(strict_types=1);
+
+namespace Cycle\Database\Tests\Driver\Postgres;
+
+/**
+ * @group driver
+ * @group driver-postgres
+ */
+class SelectWithJoinQueryTest extends \Cycle\Database\Tests\SelectWithJoinQueryTest
+{
+    const DRIVER = "postgres";
+}
